+++ lib/logic/cart.dart
@@ -4,7 +4,7 @@ import 'package:charisma_of_india/customObjects/menu.dart';
 
 class Cart {
   static  List<Menu> _cart = [];
-  static int calculatedTotal = 0;
+  static double calculatedTotal = 0;
 
   //add items to the cart
   static void addToCart(Menu menu) {
@@ -22,26 +22,26 @@ class Cart {
   }
 
   // get the total cost of all menu object items in the list
-  static int getTotal() {
+  static double getTotal() {
     double total = 0;
     for (int k =0; k< _cart.length; k++) {
       total += _cart[k].price;
     }
-    return total.toInt();
+    return total;
   }
 
   // get the taxes calculation based of the menu items in the cart list
-  static int getTaxes(){
+  static double getTaxes(){
     double total = 0;
     for (int k =0; k< _cart.length; k++) {
       total += _cart[k].price;
     }
     double result = 0.13 * total;
-    return result.toInt();
+    return result;
   }
 
   // gets the total between the taxes and any other service charges
-  static int getTotalAmount(int total, int taxes, int delivery){
+  static double getTotalAmount(double total, double taxes, double delivery){
     calculatedTotal = total + taxes + delivery;
     return calculatedTotal;
   }
